+++ driver/identity.go
@@ -26,6 +26,13 @@ func (d *QuobyteDriver) GetPluginCapabilities(ctx context.Context, req *csi.GetP
 					},
 				},
 			},
+			{
+				Type: &csi.PluginCapability_VolumeExpansion_{
+					VolumeExpansion: &csi.PluginCapability_VolumeExpansion{
+						Type: csi.PluginCapability_VolumeExpansion_ONLINE,
+					},
+				},
+			},
 		},
 	}
 	return resp, nil
